Move flood tests to CVE directory
