+++ tests/stress/test_stress.py
@@ -6,11 +6,9 @@ import os
 import time
 from pathlib import Path
 
-from framework.deproxy.deproxy_message import HttpMessage
 from framework.helpers import dmesg, networker, remote, tf_cfg
 from framework.services import tempesta
 from framework.test_suite import marks, tester
-from tests.frang.frang_test_case import FrangTestCase
 
 __author__ = "Tempesta Technologies, Inc."
 __copyright__ = "Copyright (C) 2022-2026 Tempesta Technologies, Inc."
@@ -846,159 +844,3 @@ class TestContinuationFlood(tester.TempestaTest):
         await self.wait_while_busy(client)
         client.stop()
         self.assertEqual(0, client.returncode)
-
-
-class TestRequestsUnderCtrlFrameFlood(FrangTestCase):
-    """
-    Test ability to handle requests from the client
-    under control frames frame flood.
-    Also check that there is no kernel BUGS and WARNINGs
-    under flood.
-    """
-
-    backends = [
-        {
-            "id": "deproxy",
-            "type": "deproxy",
-            "port": "8000",
-            "response": "static",
-            "response_content": (
-                "HTTP/1.1 200 OK\r\n"
-                + f"Date: {HttpMessage.date_time_string()}\r\n"
-                + "Server: debian\r\n"
-                + "Content-Length: 0\r\n\r\n"
-            ),
-        }
-    ]
-
-    tempesta = {
-        "config": """
-        listen 443 proto=h2;
-
-        server ${server_ip}:8000;
-
-        tls_certificate ${tempesta_workdir}/tempesta.crt;
-        tls_certificate_key ${tempesta_workdir}/tempesta.key;
-        tls_match_any_server_name;
-        cache 0;
-    """
-    }
-
-    clients = [
-        {
-            "id": "ctrl_frames_flood",
-            "type": "external",
-            "binary": "ctrl_frames_flood",
-            "ssl": True,
-            "cmd_args": "",
-        },
-    ]
-
-    async def _test(self, cmd_args):
-        await self.start_all_services(client=False)
-        flood_client = self.get_client("ctrl_frames_flood")
-        flood_client.options = [cmd_args % tf_cfg.cfg.get("Tempesta", "ip")]
-        flood_client.start()
-        await self.wait_while_busy(flood_client)
-        flood_client.stop()
-
-    def _check_ping_frame_exceeded(self):
-        tempesta = self.get_tempesta()
-        tempesta.get_stats()
-        self.assertEqual(tempesta.stats.cl_ping_frame_exceeded, 100)
-
-    def _check_prio_frame_exceeded(self):
-        tempesta = self.get_tempesta()
-        tempesta.get_stats()
-        self.assertEqual(tempesta.stats.cl_priority_frame_exceeded, 100)
-
-    def _check_settings_frame_exceeded(self):
-        tempesta = self.get_tempesta()
-        stats = tempesta.get_stats()
-        self.assertEqual(tempesta.stats.cl_settings_frame_exceeded, 100)
-
-    def _check_wnd_update_frame_exceeded(self):
-        tempesta = self.get_tempesta()
-        stats = tempesta.get_stats()
-        self.assertEqual(tempesta.stats.cl_wnd_update_frame_exceeded, 100)
-
-    def _check_rst_frame_exceeded(self):
-        tempesta = self.get_tempesta()
-        stats = tempesta.get_stats()
-        self.assertEqual(tempesta.stats.cl_rst_frame_exceeded, 100)
-
-    @marks.Parameterize.expand(
-        [
-            marks.Param(
-                name="PingFlood",
-                cmd_args=f"-address %s:443 -threads 4 -connections 100 -ctrl_frame_type ping_frame -frame_count 100000",
-                check_func=_check_ping_frame_exceeded,
-            ),
-            marks.Param(
-                name="SettingsFlood",
-                cmd_args=f"-address %s:443 -threads 4 -connections 100 -ctrl_frame_type settings_frame -frame_count 100000",
-                check_func=_check_settings_frame_exceeded,
-            ),
-            marks.Param(
-                name="WndUpdateFlood",
-                cmd_args=f"-address %s:443 -threads 4 -connections 100 -ctrl_frame_type window_update -frame_count 100000",
-                check_func=_check_wnd_update_frame_exceeded,
-            ),
-            marks.Param(
-                name="RstFloodByWndUpdate",
-                cmd_args=f"-address %s:443 -threads 4 -connections 100 -ctrl_frame_type rapid_reset -rapid_reset_type window_update -frame_count 100000",
-                check_func=_check_rst_frame_exceeded,
-            ),
-            marks.Param(
-                name="RstFloodByPriority",
-                cmd_args=f"-address %s:443 -threads 4 -connections 100 -ctrl_frame_type rapid_reset -rapid_reset_type priority -frame_count 100000",
-                check_func=_check_rst_frame_exceeded,
-            ),
-            marks.Param(
-                name="RstFloodByRst",
-                cmd_args=f"-address %s:443 -threads 4 -connections 100 -ctrl_frame_type rapid_reset -rapid_reset_type rst -frame_count 100000",
-                check_func=_check_rst_frame_exceeded,
-            ),
-            marks.Param(
-                name="RstFloodByRstBatch",
-                cmd_args=f"-address %s:443 -threads 4 -connections 100 -ctrl_frame_type rapid_reset -rapid_reset_type batch -frame_count 100000",
-                check_func=_check_rst_frame_exceeded,
-            ),
-            marks.Param(
-                name="RstByHeadersMaxConcurrentStreamsExceeded",
-                cmd_args=f"-address %s:443 -threads 4 -connections 100 -ctrl_frame_type rapid_reset -rapid_reset_type headers_by_max_streams_exceeded -frame_count 100000",
-                check_func=_check_rst_frame_exceeded,
-            ),
-            marks.Param(
-                name="RstByHeadersInvalidDependency",
-                cmd_args=f"-address %s:443 -threads 4 -connections 100 -ctrl_frame_type rapid_reset -rapid_reset_type headers_by_invalid_dependency -frame_count 100000",
-                check_func=_check_rst_frame_exceeded,
-            ),
-            marks.Param(
-                name="RstByIncorrectFrameType",
-                cmd_args=f"-address %s:443 -threads 4 -connections 100 -ctrl_frame_type rapid_reset -rapid_reset_type incorrect_frame_type -frame_count 100000",
-                check_func=_check_rst_frame_exceeded,
-            ),
-            marks.Param(
-                name="RstByIncorrectHeader",
-                cmd_args=f"-address %s:443 -threads 4 -connections 100 -ctrl_frame_type rapid_reset -rapid_reset_type incorrect_header -frame_count 100000",
-                check_func=_check_rst_frame_exceeded,
-            ),
-        ]
-    )
-    @dmesg.unlimited_rate_on_tempesta_node
-    async def test(self, name, cmd_args, check_func):
-        server = self.get_server("deproxy")
-        response_body = 2000 * "a"
-        server.set_response(
-            "HTTP/1.1 200 OK\r\n"
-            + "Server: Debian\r\n"
-            + f"Date: {HttpMessage.date_time_string()}\r\n"
-            + f"Content-Length: {len(response_body)}\r\n\r\n"
-            + response_body
-        )
-        await self._test(cmd_args)
-        check_func(self)
-
-
-# vim: tabstop=8 expandtab shiftwidth=4 softtabstop=4
