+++ tests/tests_disabled.json
@@ -191,22 +191,22 @@
         },
         {
             "name": "tests.stress.test_stress.TestStressNoCacheMTU80",
-            "reason": "Disabled by test issue #817"
+            "reason": "Disabled by test issue #"
         },
         {
             "name": "tests.http2_general.test_h2_frame.TestPostponedFrames",
             "reason": "Disabled by test issue #862"
         },
-        {
+              {
             "name": "tests.server_connections.test_close_dead_connections.TestFinishH2StreamsByClient.test_drop_server_connection_for_goaway",
             "reason": "Disabled by issue #2626"
         },
         {
-            "name": "tests.cve.test_cve_2019.TestSlowRead.test_cve_2019_9511",
+            "name": "tests.cve.test_cve.TestSlowRead.test_cve_2019_9511",
             "reason": "Disabled by issue #2627"
         },
         {
-            "name": "tests.cve.test_cve_2019.TestSlowRead.test_cve_2019_9517",
+            "name": "tests.cve.test_cve.TestSlowRead.test_cve_2019_9517",
             "reason": "Disabled by issue #1715"
         }
     ]
