+++ tests/tests_disabled_dbgkernel.json
@@ -22,7 +22,7 @@
             "reason" : "the tests did not work on the first run"
         },
         {
-            "name": "tests.stress.test_stress.TestContinuationFlood",
+            "name": "tests.cve.test_cve.TestHttp2FrameFlood.test_cve_2024_2758",
             "reason" : "the tests did not work on the first run"
         },
         {
