+++ tests/tests_disabled_remote.json
@@ -107,15 +107,15 @@
         },
         {
             "name": "tests.stress.test_stress.H2LoadStressMTU80",
-            "reason": "Disabled by test issue #817"
+            "reason": "Disabled by test issue #816"
         },
         {
             "name": "tests.stress.test_stress.TlsWrkStressMTU80",
-            "reason": "Disabled by test issue #817"
+            "reason": "Disabled by test issue #816"
         },
         {
             "name": "tests.stress.test_stress.WrkStressMTU80",
-            "reason": "Disabled by test issue #817"
+            "reason": "Disabled by test issue #816"
         }
     ]
 }
